+++ change_log.md
@@ -1,3 +1,6 @@
+# 2.3.4
+- fixed unitialized editor pointer
+- made title a button which leads to TheWaveWarden homepage
 # 2.3.3
 - added CLAP plugin support
 - added native M1 Mac support
@@ -5,6 +8,7 @@
 - switched build system to CMake
 - made WinInstaller suitable for systems not mounted on C:
 - fixed symbol name clashes for "Knob" and "Filter"
+- fixed issue where renaming a soundbank would select the bank above
 # 2.3.2
 - fixed issue where multiple instances of reverb would produce noise
 # 2.3.1